Main module - client details (wip).
git-svn-id: http://svn.apollosoft.net.au:18080/svn/myplanner/trunk@59 65940d0d-0698-491e-b608-6f35d31c8762

+++ core/src/main/java/com/argus/financials/domain/hibernate/AbstractAuditable.java
@@ -24,7 +24,9 @@ public abstract class AbstractAuditable<T> extends AbstractBase<T>
     private Date dateModified;
 
     @Transient
-    private Long version;
+    //@Version
+    //@Column(name = "")
+    private Integer version;
 
     /**
      * @return the dateCreated
@@ -61,7 +63,7 @@ public abstract class AbstractAuditable<T> extends AbstractBase<T>
     /**
      * @return the version
      */
-    public Long getVersion()
+    public Integer getVersion()
     {
         return version;
     }
@@ -69,7 +71,7 @@ public abstract class AbstractAuditable<T> extends AbstractBase<T>
     /**
      * @param version the version to set
      */
-    public void setVersion(Long version)
+    public void setVersion(Integer version)
     {
         this.version = version;
     }
